+++ R/performance_mse.R
@@ -18,8 +18,7 @@
 #' data(mtcars)
 #' m <- lm(mpg ~ hp + gear, data = mtcars)
 #' performance_mse(m)
-#' @importFrom insight print_color
-#' @importFrom stats residuals predict
+#' @importFrom insight get_residuals
 #' @export
 performance_mse <- function(model, ...) {
   UseMethod("performance_mse")
@@ -35,11 +34,8 @@ mse <- performance_mse
 performance_mse.default <- function(model, verbose = TRUE, ...) {
   res <- tryCatch(
     {
-      pred <- stats::predict(model, type = "response")
-      observed <- .factor_to_numeric(insight::get_response(model))
-      observed - pred
-    },
-    error = function(e) {
+      insight::get_residuals(model, verbose = verbose, type = "response", ...)
+    }, error = function(e) {
       NULL
     }
   )
@@ -47,26 +43,22 @@ performance_mse.default <- function(model, verbose = TRUE, ...) {
   if (is.null(res)) {
     res <- tryCatch(
       {
-        if (inherits(model, c("vgam", "vglm"))) {
-          model@residuals
-        } else if (inherits(model, "coxph")) {
-          stats::residuals(model)
-        } else {
-          stats::residuals(model, type = "response")
+        def_res <- insight::get_residuals(model, verbose = verbose, ...)
+        if (verbose) {
+          warning("Response residuals not available to calculate mean square error. (R)MSE is probably not reliable.", call. = FALSE)
         }
-      },
-      error = function(e) {
+        def_res
+      }, error = function(e) {
         NULL
       }
     )
   }
 
   if (is.null(res) || all(is.na(res))) {
-    if (verbose) insight::print_color("Can't extract residuals from model.\n", "red")
     return(NA)
   }
 
-  mean(res^2, na.rm = T)
+  mean(res^2, na.rm = TRUE)
 }
 
 
@@ -102,29 +94,3 @@ performance_mse.betaor <- performance_mse.logitor
 
 #' @export
 performance_mse.betamfx <- performance_mse.logitor
-
-
-
-
-# other models --------------------------
-
-#' @export
-performance_mse.slm <- function(model, verbose = TRUE, ...) {
-  res <- tryCatch(
-    {
-      junk <- utils::capture.output(pred <- stats::predict(model, type = "response"))
-      observed <- .factor_to_numeric(insight::get_response(model))
-      observed - pred
-    },
-    error = function(e) {
-      NULL
-    }
-  )
-
-  if (is.null(res) || all(is.na(res))) {
-    if (verbose) insight::print_color("Can't extract residuals from model.\n", "red")
-    return(NA)
-  }
-
-  mean(res^2, na.rm = T)
-}